⚡ Optimize re-renders and fix initialization order
- Relocated the `download` utility function to the module scope for a stable identity.
- Wrapped all event handlers passed to child panels in `useCallback` to prevent unnecessary re-renders.
- Fixed a potential `ReferenceError` by moving the `showToast` definition to the top of the component, ensuring it is initialized before being used in other `useCallback` hooks.
- Memoized `EditorPanel`, `PreviewPanel`, and `UploadPanel` with `React.memo` to skip rendering when their props (now stable) don't change.

These changes significantly improve the responsiveness of the editor by ensuring that only the necessary parts of the UI update as the user types.
