+++ components/MonacoEditor.tsx
@@ -1,29 +1,143 @@
 "use client";
 
-import React from "react";
-import dynamic from "next/dynamic";
-import { MonacoEditorProps } from "./types";
+import React, { useRef, useEffect, useState } from "react";
+import { Editor } from "@monaco-editor/react";
 
-const MonacoEditorClient = dynamic(() => import("./MonacoEditorClient"), {
-  ssr: false,
-  loading: () => (
-    <div className="monaco-loading">
-      <div className="loading-text">Loading editor...</div>
-    </div>
-  ),
-});
+interface MonacoEditorProps {
+  value: string;
+  onChange: (value: string) => void;
+  className?: string;
+}
 
 export default function MonacoEditor({
   value,
   onChange,
   className = ""
 }: MonacoEditorProps) {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const editorRef = useRef<any>(null);
+  const [theme, setTheme] = useState('vs-dark');
+  const [isClient, setIsClient] = useState(false);
+
+  // Check if we're on the client side
+  useEffect(() => {
+    // eslint-disable-next-line react-hooks/set-state-in-effect
+    setIsClient(true);
+  }, []);
+
+  // Detect theme from HTML class
+  useEffect(() => {
+    if (!isClient) return;
+
+    const updateTheme = () => {
+      const isDark = document.documentElement.classList.contains('dark') ||
+                    !document.documentElement.classList.contains('light');
+      setTheme(isDark ? 'vs-dark' : 'vs');
+    };
+
+    updateTheme();
+
+    // Watch for theme changes
+    const observer = new MutationObserver(updateTheme);
+    observer.observe(document.documentElement, {
+      attributes: true,
+      attributeFilter: ['class']
+    });
+
+    return () => observer.disconnect();
+  }, [isClient]);
+
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const handleEditorDidMount = (editor: any) => {
+    editorRef.current = editor;
+
+    // Configure editor options
+    editor.updateOptions({
+      minimap: { enabled: false },
+      scrollBeyondLastLine: false,
+      wordWrap: 'on',
+      lineNumbers: 'off',
+      glyphMargin: false,
+      folding: false,
+      lineDecorationsWidth: 0,
+      lineNumbersMinChars: 0,
+      renderLineHighlight: 'none',
+      selectOnLineNumbers: false,
+      roundedSelection: false,
+      readOnly: false,
+      cursorStyle: 'line',
+      automaticLayout: true,
+      fontSize: 14,
+      fontFamily: 'ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace',
+      lineHeight: 24,
+      padding: { top: 16, bottom: 16 },
+      scrollbar: {
+        vertical: 'auto',
+        horizontal: 'auto',
+        useShadows: false,
+        verticalHasArrows: false,
+        horizontalHasArrows: false,
+      },
+    });
+
+    // Language is already set via defaultLanguage prop
+  };
+
+  const handleEditorChange = (value: string | undefined) => {
+    if (value !== undefined) {
+      onChange(value);
+    }
+  };
+
+
+  // Show loading state during SSR
+  if (!isClient) {
+    return (
+      <div className={`monaco-editor-container ${className}`}>
+        <div className="monaco-loading">
+          <div className="loading-text">Loading editor...</div>
+        </div>
+      </div>
+    );
+  }
+
   return (
     <div className={`monaco-editor-container ${className}`}>
-      <MonacoEditorClient
+      <Editor
+        height="100%"
+        width="100%"
+        defaultLanguage="markdown"
         value={value}
-        onChange={onChange}
-        className={className}
+        onChange={handleEditorChange}
+        onMount={handleEditorDidMount}
+        theme={theme}
+        options={{
+          minimap: { enabled: false },
+          scrollBeyondLastLine: false,
+          wordWrap: 'on',
+          lineNumbers: 'off',
+          glyphMargin: false,
+          folding: false,
+          lineDecorationsWidth: 0,
+          lineNumbersMinChars: 0,
+          renderLineHighlight: 'none',
+          selectOnLineNumbers: false,
+          roundedSelection: false,
+          readOnly: false,
+          cursorStyle: 'line',
+          automaticLayout: true,
+          fontSize: 14,
+          fontFamily: 'ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace',
+          lineHeight: 24,
+          padding: { top: 16, bottom: 16 },
+          scrollbar: {
+            vertical: 'auto',
+            horizontal: 'auto',
+            useShadows: false,
+            verticalHasArrows: false,
+            horizontalHasArrows: false,
+          },
+        }}
       />
     </div>
   );
