+++ hooks/useIsMobile.ts
@@ -7,29 +7,19 @@ export function useIsMobile(breakpoint: number = 768) {
   const [isLoading, setIsLoading] = useState(true);
 
   useEffect(() => {
-    let timeoutId: ReturnType<typeof setTimeout>;
-
     const checkIsMobile = () => {
       setIsMobile(window.innerWidth < breakpoint);
       setIsLoading(false);
-    };
-
-    const handleResize = () => {
-      clearTimeout(timeoutId);
-      timeoutId = setTimeout(checkIsMobile, 100);
     };
 
     // Check on mount
     checkIsMobile();
 
     // Add event listener for resize
-    window.addEventListener("resize", handleResize);
+    window.addEventListener("resize", checkIsMobile);
 
     // Cleanup
-    return () => {
-      window.removeEventListener("resize", handleResize);
-      clearTimeout(timeoutId);
-    };
+    return () => window.removeEventListener("resize", checkIsMobile);
   }, [breakpoint]);
 
   return { isMobile, isLoading };
